// frontend/src/api/generated/reservations/reservations.zod.ts
export const PostApiReservationsWithJsonBody = zod.object({
  "factoryId": zod.number(),
  "startTime": zod.iso.datetime({}),
  "endTime": zod.iso.datetime({}),
  "personIds": zod.array(zod.number())
})
// ...
export const PutApiReservationsIdWithJsonBody = zod.object({
  "factoryId": zod.number(),
  "startTime": zod.iso.datetime({}),
  "endTime": zod.iso.datetime({}),
  "personIds": zod.array(zod.number())
})